@@ -1,18 +1,18 @@
 name: shared_preferences_platform_interface
 description: A common platform interface for the shared_preferences plugin.
 homepage: https://github.com/flutter/plugins/tree/master/packages/shared_preferences/shared_preferences_platform_interface
-version: 1.0.4
+version: 2.0.0-nullsafety
 
 dependencies:
-  meta: ^1.0.4
+  meta: ^1.3.0-nullsafety.6
   flutter:
     sdk: flutter
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0-nullsafety.3
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.8 <2.0.0"
